requirements.txt
gradio
numpy
Pillow
git+https://github.com/facebookresearch/segment-anything.git
torch
torchvision
